Fix bpm on mac (#1125)
